% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/resp.R
\docType{data}
\name{aif}
\alias{aif}
\title{An example input function}
\format{
A vector with 30 numbers

A vector with 30 numbers

A vector with 30 numbers

A vector with 30 numbers
}
\usage{
aif

aif

aif

aif
}
\description{
An example input function

An example input function

An example input function

An example input function
}
\keyword{datasets}
