% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{api_base}
\alias{api_base}
\title{Retrieve the API base url}
\usage{
api_base(name = "ops", segments = list(NULL, "search/granules.umm_json")[[1]])
}
\arguments{
\item{name}{char, one of 'ops', 'uat' or 'sit'}

\item{segments}{NULL or character to append to the URL}
}
\value{
character URL
}
\description{
Retrieve the API base url
}
