% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{as_xml}
\alias{as_xml}
\title{Given a credentials list, convert to an xml node}
\usage{
as_xml(x = read_credentials())
}
\arguments{
\item{x}{list of credentials from \code{\link{read_credentials}}}
}
\value{
xml node
}
\description{
Given a credentials list, convert to an xml node
}
