% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{as_xml_string}
\alias{as_xml_string}
\title{Given a credentials list, convert to an xml string}
\usage{
as_xml_string(x = read_credentials())
}
\arguments{
\item{x}{list of credentials from \code{\link{read_credentials}}}
}
\value{
character, single element of XML
}
\description{
Given a credentials list, convert to an xml string
}
