% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullseye.R
\name{bullseye}
\alias{bullseye}
\title{Bullseye plot for cmr}
\usage{
bullseye(werte, zlim = NULL, reverse = TRUE, text = TRUE, cex = 2)
}
\arguments{
\item{werte}{}

\item{zlim}{}

\item{reverse}{}

\item{text}{}
}
\value{
plot
}
\description{
Bullseye plot for cmr
}

