% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmr_local.R
\name{cmr.local}
\alias{cmr.local}
\title{Title Spline analysis of cardiovascular magnetic resonance imaging}
\usage{
cmr.local(data, mask, input, quantiles = c(0.25, 0.75))
}
\arguments{
\item{data}{3d array of CMR signal}

\item{mask}{2d array of mask. Voxel with 0 or FALSE will be ommited from anaysis}

\item{input}{input function}

\item{quantiles}{quantiles used for credible intervall, default: c(0.25, 0.75)}
}
\value{
list of mbf (point estimation) and ci (credible interval)
}
\description{
Title Spline analysis of cardiovascular magnetic resonance imaging
}
\examples{
 library(cmr)
 data(sim)
 local.mbf=array(NA,c(30,30,3))
 local.ci=array(NA,c(30,30,3))
 for (i in 1:3)
 {  mask=array(NA,c(30,30))
 mask[data.data[,,i,1]!=0]=1
 temp=cmr.local(data.data[,,i,],mask,aif)
 local.mbf[,,i]=t(as.matrix(temp$mbf))
 local.ci[,,i]=t(as.matrix(temp$ci))
 }
par(mfrow=c(2,1))
imageMBF(resp,zlim=c(0,5))
imageMBF(local.mbf,zlim=c(0,5))
}

