% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrdata_sim}
\alias{cmrdata_sim}
\title{Simulated data for CMR package.}
\format{
A 4D array, 30x30 pixels for 3 slices at 30 time points.
}
\usage{
cmrdata_sim
}
\description{
This data set is provided as example for the usage of the cmR package. cmrdata_sim is a simulated CMR image.
}
\keyword{datasets}
