% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmrdata_sim.R
\docType{data}
\name{cmrsim}
\alias{cmrsim}
\alias{cmrdata_sim}
\title{Simulated data for CMR package}
\format{
An object of class \code{array} of dimension 30 x 30 x 3 x 30.
}
\usage{
data(cmrsim)
}
\description{
Simulated data for CMR package
}
\author{
Volker Schmid \email{volker.schmid@lmu.de}
}
\keyword{cmr}
\keyword{datasets}
\keyword{example}
