\name{cmrSpace-package}
\alias{cmrSpace-package}
\alias{cmrSpace}
\docType{package}
\title{
\packageTitle{cmrSpace}
}
\description{
\packageDescription{cmrSpace}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cmrSpace}
\packageIndices{cmrSpace}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cmrSpace}

Maintainer: \packageMaintainer{cmrSpace}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
