% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/resp.R
\docType{data}
\name{data}
\alias{data}
\title{CMR data example}
\format{
An array with dimension 30x30x3x30 (30x30 image, 3 slices, 30 timepoints)

An array with dimension 30x30x3x30 (30x30 image, 3 slices, 30 timepoints)
}
\usage{
data

data
}
\description{
CMR data example

CMR data example
}
\keyword{datasets}
