% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{extract_collection}
\alias{extract_collection}
\title{Extract search results into a table of links}
\usage{
extract_collection(
  x = search_collection(form = "response"),
  description_pattern = c("all", "^.*OPeNDAP")[1],
  ...
)
}
\arguments{
\item{x}{\code{\link{httr}{response}} object}

\item{description_pattern}{character, one or more regular expressions with which
to filter the 'Description' variable.  Or "all" to skip filtering.}

\item{...}{extra arguments for \code{\link[base]{grepl}} such as \code{fixed}}
}
\value{
tibble, possibly empty
\itemize{
  \item{URL, the url}
  \item{Type, describes the type of protocol for access}
  \item{Description, a verbose description}
  \item{Subtype, not really sure what this is}
}
}
\description{
Extract search results into a table of links
}
