% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_token}
\alias{get_token}
\title{Get a token}
\usage{
get_token(x = read_credentials(), delete = has_token(x), ...)
}
\arguments{
\item{x}{credentials list}

\item{delete}{logical if TRUE delete exisiting token (if it exists) first}

\item{...}{other arguments for \code{\link{token_create}}}
}
\value{
updated credentials list
}
\description{
Get a token
}
\seealso{
\href{https://wiki.earthdata.nasa.gov/display/CMR/CMR+Client+Partner+User+Guide#CMRClientPartnerUserGuide-Chapter2:GettingStarted}{CMR guide}
}
