% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{has_token}
\alias{has_token}
\title{Determine is credentials has a non-empty token}
\usage{
has_token(x = read_credentials())
}
\arguments{
\item{x}{credentials list}
}
\value{
logical, TRUE is non-empty token exists in credentials
}
\description{
Determine is credentials has a non-empty token
}
