% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_sim}
\alias{input_sim}
\title{Simulated data for CMR package.}
\format{
Vector for 30 time points.
}
\usage{
input_sim
}
\description{
This data set is provided as example for the usage of the cmR package. input_sim is the simulated input function.
}
\keyword{datasets}
