% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxresp_sim.R
\docType{data}
\name{maxresp_sim}
\alias{maxresp_sim}
\title{True maximum response function in example}
\format{
An array with dimension 30x30x3 (30x30 image, 3 slices)
}
\usage{
maxresp_sim
}
\description{
True maximum response function in example
}
\keyword{datasets}
