% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{print.cmr_credentials}
\alias{print.cmr_credentials}
\title{Print credentials by default redacted (for README)}
\usage{
\method{print}{cmr_credentials}(x, redact = TRUE, ...)
}
\arguments{
\item{x}{credentials object}

\item{redact}{logical, by default redact sensitive information}

\item{...}{other arguments for the print method}
}
\description{
Print credentials by default redacted (for README)
}
