% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobullseye.R
\name{pseudobullseye}
\alias{pseudobullseye}
\title{Pseudo bullseye plot}
\usage{
pseudobullseye(x, lim, legend = FALSE, text = TRUE, reverse = FALSE,
  center = FALSE, cex = 2)
}
\arguments{
\item{x}{3D array}

\item{lim}{limits of x values}

\item{legend}{boolean, add legend?}

\item{text}{boolean, should text legend be added?}

\item{reverse}{boolean, reverse colors?}

\item{center}{boolean, should x be centered before plotting}

\item{cex}{cex for text legend}
}
\value{
plots
}
\description{
Pseudo bullseye plot
}

