% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{read_credentials}
\alias{read_credentials}
\title{Read a credentials for EarthData}
\usage{
read_credentials(filename = "~/.earthdata")
}
\arguments{
\item{filename}{char, the name of the credentials file}
}
\value{
named list of credentials
}
\description{
Credential files should have three lines
username: <user_name>
password: <password>
token: <token_value>
}
