% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/resp.R
\docType{data}
\name{resp}
\alias{resp}
\title{True maximum response function in example}
\format{
An array with dimension 30x30x3 (30x30 image, 3 slices)

An array with dimension 30x30x3 (30x30 image, 3 slices)
}
\usage{
resp

resp
}
\description{
True maximum response function in example

True maximum response function in example
}
\keyword{datasets}
