% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{search_collection}
\alias{search_collection}
\title{Search a collection}
\usage{
search_collection(
  id = "C1996881146-POCLOUD",
  times = c("2019-01-01T10:00:00Z", "2019-02-01T00:00:00Z"),
  base_url = api_base(name = "ops", segments = "search/granules.umm_json"),
  description_pattern = "^.*OPeNDAP",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{id}{character, 'collection_concept_id'}

\item{times}{2 element character (or POSIXct) start and stop search bounds or
NULL to skip}

\item{base_url}{character, the base URL for searching}

\item{description_pattern}{character, one or more regular expressions with which
to filter the 'Description' variable.  Or "all" to skip filtering.}

\item{verbose}{logical, if TRUE print the URL}

\item{...}{extra arguments for \code{\link[base]{grepl}} such as \code{fixed}}
}
\value{
tibble of related URLs and descriptions
\itemize{
\item{URL, the url}
\item{Type, describes the type of protocol for access}
\item{Description, a verbose description}
\item{Subtype, not really sure what this is}
}
}
\description{
Search a collection
}
