% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{token_create}
\alias{token_create}
\title{Create a token}
\usage{
token_create(
  x = read_credentials(),
  uri = api_base("ops", segments = "api/users/token")
)
}
\arguments{
\item{x}{credentials list}

\item{uri}{character, the URL}
}
\value{
token or NULL if fails
}
\description{
Create a token
}
