% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{write_credentials}
\alias{write_credentials}
\title{Write a credentials for EarthData}
\usage{
write_credentials(x, filename = "~/.earthdata")
}
\arguments{
\item{x}{named list of credentials}

\item{filename}{char, the name of the credentials file}
}
\value{
the input list
}
\description{
Credential files should have three lines
username: <user_name>
password: <password>
token: <token_value>
}
