% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullseye.R
\name{bullseye}
\alias{bullseye}
\title{Bullseye plot}
\usage{
bullseye(x, lim = NULL, reverse = TRUE, legend = TRUE, text = TRUE, cex = 1)
}
\arguments{
\item{x}{vector of length 16}

\item{lim}{limits of x values}

\item{reverse}{boolean, reverse colors?}

\item{legend}{boolean, add legend?}

\item{text}{boolean, should text legend be added?}

\item{cex}{cex for text legend}
}
\value{
plot
}
\description{
Bullseye plot
}
\examples{
  bullseye(1:16)
}
