% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmrsim_data.R
\docType{data}
\name{cmrsim}
\alias{cmrsim}
\alias{aif}
\title{Simulated data for CMR package}
\format{
An object of class \code{numeric} of length 30.
}
\usage{
data(cmrsim)
}
\description{
Simulated data for CMR package
}
\author{
Volker Schmid \email{volker.schmid@lmu.de}
}
\keyword{cmr}
\keyword{datasets}
\keyword{example}
