% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maxresp_sim}
\alias{maxresp_sim}
\title{Simulated data for CMR package.}
\format{
A 3D array, 30x30 pixels for 3 slices.
}
\usage{
maxresp_sim
}
\description{
This data set is provided as example for the usage of the cmR package. maxresp_sim is the true maximum response used in the simulation.
}
\keyword{datasets}
