% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{search_collection}
\alias{search_collection}
\title{Search a collection}
\usage{
search_collection(
  id = "C1996881146-POCLOUD",
  times = c("2019-01-01T10:00:00Z", "2019-02-01T00:00:00Z"),
  base_url = api_base(name = "ops", segments = "search/granules.umm_json"),
  verbose = FALSE,
  form = c("response", "table", "url")[2],
  ...
)
}
\arguments{
\item{id}{character, 'collection_concept_id'}

\item{times}{2 element character (or POSIXct) start and stop search bounds or
NULL to skip}

\item{base_url}{character, the base URL for searching}

\item{verbose}{logical, if TRUE print the URL}

\item{form}{character, "response" to return a \code{\link{httr}{response}} object,
"url" to retrieve the charcater URL, or 
"table" to extract a table of links}

\item{...}{extra arguments for \code{\link{extract_collection}}}
}
\value{
\code{\link{httr}{response}} object, a tibble as per \code{\link{extract_collection}}
}
\description{
Search a collection
}
