% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{token_create}
\alias{token_create}
\title{Create a token}
\usage{
token_create(
  x = read_credentials(),
  verbose = FALSE,
  uri = api_base("ops", segments = "legacy-services/rest/tokens")
)
}
\arguments{
\item{x}{credentials list}

\item{verbose}{logical, if TRUE echo the command string}

\item{uri}{character, the URL}

\item{use}{character one of 'curl' or 'httr'}
}
\value{
character token or NULL if fails
}
\description{
Create a token
}
\seealso{
\href{https://wiki.earthdata.nasa.gov/display/CMR/CMR+Client+Partner+User+Guide#CMRClientPartnerUserGuide-Chapter2:GettingStarted}{CMR guide}
}
